+++ src/main/java/org/broadinstitute/utils/MathUtils.java
@@ -6,17 +6,17 @@ package org.broadinstitute.utils;
  */
 public class MathUtils {
 
-    private static final double[] jacobianLogTable;
-    private static final double JACOBIAN_LOG_TABLE_STEP = 0.001;
-    private static final double MAX_JACOBIAN_TOLERANCE = 8.0;
-    private static final double JACOBIAN_LOG_TABLE_INV_STEP = 1.0 / 0.001;
+    private static final float[] jacobianLogTable;
+    private static final float JACOBIAN_LOG_TABLE_STEP = 0.001f;
+    private static final float MAX_JACOBIAN_TOLERANCE = 8.f;
+    private static final float JACOBIAN_LOG_TABLE_INV_STEP = 1.f / 0.001f;
     private static final int JACOBIAN_LOG_TABLE_SIZE = (int) (MAX_JACOBIAN_TOLERANCE / JACOBIAN_LOG_TABLE_STEP) + 1;
 
     static {
-        jacobianLogTable = new double[JACOBIAN_LOG_TABLE_SIZE];
+        jacobianLogTable = new float[JACOBIAN_LOG_TABLE_SIZE];
 
         for (int k = 0; k < JACOBIAN_LOG_TABLE_SIZE; k++) {
-            jacobianLogTable[k] = Math.log10(1.0 + Math.pow(10.0, -((double) k) * JACOBIAN_LOG_TABLE_STEP));
+            jacobianLogTable[k] = (float) Math.log10(1.0 + Math.pow(10.0, -((float) k) * JACOBIAN_LOG_TABLE_STEP));
 
         }
     }
@@ -28,7 +28,7 @@ public class MathUtils {
      *               -Infinity values, as log10(0.0) == -Infinity.
      * @return true if result is really well formed
      */
-    public static boolean goodLog10Probability(final double result) {
+    public static boolean goodLog10Probability(final float result) {
         return goodLog10Probability(result, true);
     }
 
@@ -39,18 +39,18 @@ public class MathUtils {
      * @param allowNegativeInfinity should we consider a -Infinity value ok?
      * @return true if result is really well formed
      */
-    public static boolean goodLog10Probability(final double result, final boolean allowNegativeInfinity) {
-        return result <= 0.0 && result != Double.POSITIVE_INFINITY && (allowNegativeInfinity || result != Double.NEGATIVE_INFINITY) && ! Double.isNaN(result);
+    public static boolean goodLog10Probability(final float result, final boolean allowNegativeInfinity) {
+        return result <= 0.0 && result != Float.POSITIVE_INFINITY && (allowNegativeInfinity || result != Float.NEGATIVE_INFINITY) && ! Float.isNaN(result);
     }
 
     // A fast implementation of the Math.round() method.  This method does not perform
     // under/overflow checking, so this shouldn't be used in the general case (but is fine
     // if one is already make those checks before calling in to the rounding).
-    public static int fastRound(double d) {
+    public static int fastRound(float d) {
         return (d > 0.0) ? (int) (d + 0.5d) : (int) (d - 0.5d);
     }
 
-    public static int maxElementIndex(final double[] array, final int endIndex) {
+    public static int maxElementIndex(final float[] array, final int endIndex) {
         if (array == null || array.length == 0)
             throw new IllegalArgumentException("Array cannot be null!");
 
@@ -63,42 +63,42 @@ public class MathUtils {
         return maxI;
     }
 
-    public static int maxElementIndex(final double[] array) {
+    public static int maxElementIndex(final float[] array) {
         return maxElementIndex(array, array.length);
     }
 
-    public static double arrayMax(final double[] array, final int endIndex) {
+    public static float arrayMax(final float[] array, final int endIndex) {
         return array[maxElementIndex(array, endIndex)];
     }
 
-    public static double log10sumLog10(double[] log10p, int start, int finish) {
-        double sum = 0.0;
+    public static float log10sumLog10(float[] log10p, int start, int finish) {
+        float sum = 0.f;
 
-        double maxValue = arrayMax(log10p, finish);
-        if(maxValue == Double.NEGATIVE_INFINITY)
+        float maxValue = arrayMax(log10p, finish);
+        if(maxValue == Float.NEGATIVE_INFINITY)
             return maxValue;
 
         for (int i = start; i < finish; i++) {
-            sum += Math.pow(10.0, log10p[i] - maxValue);
+            sum += (float) Math.pow(10.0, log10p[i] - maxValue);
         }
 
-        return Math.log10(sum) + maxValue;
+        return (float) Math.log10(sum) + maxValue;
     }
 
-    public static double approximateLog10SumLog10(final double[] vals) {
+    public static float approximateLog10SumLog10(final float[] vals) {
         return approximateLog10SumLog10(vals, vals.length);
     }
 
-    public static double approximateLog10SumLog10(final double[] vals, final int endIndex) {
+    public static float approximateLog10SumLog10(final float[] vals, final int endIndex) {
 
         final int maxElementIndex = MathUtils.maxElementIndex(vals, endIndex);
-        double approxSum = vals[maxElementIndex];
+        float approxSum = vals[maxElementIndex];
 
         for (int i = 0; i < endIndex; i++) {
-            if (i == maxElementIndex || vals[i] == Double.NEGATIVE_INFINITY)
+            if (i == maxElementIndex || vals[i] == Float.NEGATIVE_INFINITY)
                 continue;
 
-            final double diff = approxSum - vals[i];
+            final float diff = approxSum - vals[i];
             if (diff < MathUtils.MAX_JACOBIAN_TOLERANCE) {
                 // See notes from the 2-inout implementation below
                 final int ind = fastRound(diff * MathUtils.JACOBIAN_LOG_TABLE_INV_STEP); // hard rounding
@@ -109,26 +109,26 @@ public class MathUtils {
         return approxSum;
     }
 
-    public static double log10sumLog10(double[] log10p, int start) {
+    public static float log10sumLog10(float[] log10p, int start) {
         return log10sumLog10(log10p, start, log10p.length);
     }
 
-    public static double log10sumLog10(double[] log10values) {
+    public static float log10sumLog10(float[] log10values) {
         return log10sumLog10(log10values, 0);
     }
 
-    public static double approximateLog10SumLog10(double small, double big) {
+    public static float approximateLog10SumLog10(float small, float big) {
         // make sure small is really the smaller value
         if (small > big) {
-            final double t = big;
+            final float t = big;
             big = small;
             small = t;
         }
 
-        if (small == Double.NEGATIVE_INFINITY)
+        if (small == Float.NEGATIVE_INFINITY)
             return big;
 
-        final double diff = big - small;
+        final float diff = big - small;
         if (diff >= MAX_JACOBIAN_TOLERANCE)
             return big;
